requirements.txt
flask
openai
langchain
faiss-cpu
python-dotenv
sentence-transformers